Remove real time linting during development
While real time linting is useful, it requires a lot of extra computation. Since most of the IDEs already include an internal linter or provide a plugin for that purpose, linting each time there is an edit in the project seems an useless extra effort. It is better to compile the project as fast as possible instead of keeping it always clean. Of course, the linting process while building the project for production remains