Dedupe SCSS webpack configuration
SCSS webpack configuration was exactly the same in production and in development (that's because the minification in production is handled by `postcss.config.js`). This means that it can be safely moved inside `base.config.js`. The project will keep working in the exact same way as before, but without a useless code repetition